Cucumber-related classes to a separate package moved
