+++ src/test/java/ru/mystamps/web/tests/integration/cucumber/CucumberRunner.java
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
-package ru.mystamps.web.tests;
+package ru.mystamps.web.tests.integration.cucumber;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
